refactor(theme): finalize `decay` -> `color_scale`
This finalizes the move from `decay` to `color_scale`, mostly via a `sed
-i 's/decay/color_scale/g' src/**.rs`, but doc errors emerging from this
move were also corrected.
