+++ src/output/color_scale.rs
